% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PRS.R
\name{calculate_PRS}
\alias{calculate_PRS}
\title{Calculate individual PRS for each trait based on refined effect size estimates from PolyPred}
\usage{
calculate_PRS(K, base, pcut, G.target)
}
\arguments{
\item{K}{number of traits}

\item{base}{list of disease GWAS summary statistics of K traits from base cohort}

\item{pcut}{p-value cutoff}

\item{G.target}{genotype in target cohort}
}
\value{
a matrix of K individual PRSs
}
\description{
Calculate individual PRS for each trait based on refined effect size estimates from PolyPred
}
\details{
calculate_PRS needs disease GWAS summary statistics from K traits, assuming the effect size estimates have been refined by PolyPred after fine-mapping
}
\examples{
\donttest{
dat <- generate_pgx_data(structure = "CS", sparseness = "no",
rho_DT = c(0.5,0.5,0.5,0.5), rho_T = 0.5, rho_E = 0.5, rho_C = 0.2,
K=4, m=2000, pcausal=0.1, blocksize=100,
gamma=1, samplesize=700, h2_base=0.3, h2_target=0.3)
re <- calculate_PRS(K = 4, base = dat$base, pcut = 1, G.target = dat$target[,-c(1,2)])
head(re)
}

}
\references{
Zhai, S., Guo, B., and Shen, J., 2025. Improving multi-trait polygenic risk score prediction using fine-mapping and ensemble learning.
}
\author{
Song Zhai
}
