% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weight.R
\name{calculate_weight}
\alias{calculate_weight}
\title{Calculate weights by by performing principal component analysis (PCA) on genetic correlation matrix among traits}
\usage{
calculate_weight(K, corr, varcut)
}
\arguments{
\item{K}{number of traits}

\item{corr}{genetic correlation matrix}

\item{varcut}{variance cutoff}
}
\value{
a vector of weights
}
\description{
Calculate weights by by performing principal component analysis (PCA) on genetic correlation matrix among traits
}
\details{
Calculate weights by by performing principal component analysis (PCA) on genetic correlation matrix among traits
}
\examples{
\donttest{
dat <- generate_pgx_data(structure = "CS", sparseness = "no",
rho_DT = c(0.5,0.5,0.5,0.5), rho_T = 0.5, rho_E = 0.5, rho_C = 0.2,
K=4, m=2000, pcausal=0.1, blocksize=100,
gamma=1, samplesize=700, h2_base=0.3, h2_target=0.3)
re <- calculate_weight(K = 4, corr = dat$corr, varcut = 0.8)
hist(re$mtPRS)
}

}
\references{
Zhai, S., Guo, B., and Shen, J., 2025. Improving multi-trait polygenic risk score prediction using fine-mapping and ensemble learning.
}
\author{
Song Zhai
}
