% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pgx_data.R
\name{generate_pgx_data}
\alias{generate_pgx_data}
\title{Simulate PGx GWAS data for mtPRS analysis}
\usage{
generate_pgx_data(
  structure = "CS",
  sparseness = "no",
  rho_DT,
  rho_T,
  rho_E,
  rho_C,
  K = 4,
  m = 1000,
  pcausal = 0.1,
  blocksize = 100,
  gamma = 1,
  samplesize = 700,
  h2_base = 0.3,
  h2_target = 0.3
)
}
\arguments{
\item{structure}{genetic correlation structure, either uniformly correlated (i.e., "CS" or "AR1") or clustered correlated ("clustered")}

\item{sparseness}{effect sparsity, either "no" or "half" or "more"}

\item{rho_DT}{effect correlations between traits in base cohort and the phenotype in target cohort}

\item{rho_T}{genetic correlation among traits in base cohort}

\item{rho_E}{correlation between prognostic and predictive effects in target cohort}

\item{rho_C}{between-cluster correlation when the genetic correlation structure is clustered correlated}

\item{K}{number of traits}

\item{m}{number of SNPs}

\item{pcausal}{proportion of causal SNPs}

\item{blocksize}{LD block size (i.e., the number of SNPs in each LD block)}

\item{gamma}{prognostic-to-predictive effect size ratio}

\item{samplesize}{sample size of target cohort}

\item{h2_base}{heritability in base cohort}

\item{h2_target}{heritability in target cohort}
}
\value{
a list of disease GWAS summary statistics in base cohort, individual-level data in validation and target cohorts, genetic correlation matrix among traits, and true effect sizes in base and target cohorts
}
\description{
Simulate PGx GWAS data for mtPRS analysis
}
\details{
Simulate PGx GWAS data for mtPRS analysis
}
\examples{
\donttest{
dat <- generate_pgx_data(structure = "CS", sparseness = "no",
rho_DT = c(0.5,0.5,0.5,0.5), rho_T = 0.5, rho_E = 0.5, rho_C = 0.2,
K=4, m=2000, pcausal=0.1, blocksize=100,
gamma=1, samplesize=700, h2_base=0.3, h2_target=0.3)
plot(dat$truesize[,1])
print(dat$corr)
}

}
\references{
Zhai, S., Guo, B., and Shen, J., 2025. Improving multi-trait polygenic risk score prediction using fine-mapping and ensemble learning.
}
\author{
Song Zhai
}
