% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtPRS_FIMEL.R
\name{mtPRS_FIMEL}
\alias{mtPRS_FIMEL}
\title{Construct mtPRS by performing ensemble learning}
\usage{
mtPRS_FIMEL(
  base,
  target,
  validation,
  corr,
  pcut = 0.05,
  varcut = 0.8,
  K = 4,
  phenotype = "pgx"
)
}
\arguments{
\item{base}{list of disease GWAS summary statistics of K traits}

\item{target}{data frame in target cohort, including Y, T (if the target cohort is from a PGx study), and G}

\item{validation}{data frame from an independent validation cohort, including Y, T (if the validation cohort is from a PGx study), and G}

\item{corr}{genetic correlation matrix among K traits}

\item{pcut}{p-value cutoff for C+T method to construct individual stPRS}

\item{varcut}{variance cutoff to choose top principal components (PCs)}

\item{K}{number of traits}

\item{phenotype}{indicator of phenotype in target cohort, either "dis" or "pgx"}
}
\value{
a list of fitted model, stPRSs and mtPRSs, and mtPRS-FIMEL
}
\description{
Construct mtPRS by performing ensemble learning
}
\details{
Construct mtPRS by performing ensemble learning
}
\examples{
\donttest{
dat <- generate_pgx_data(structure = "CS", sparseness = "no",
rho_DT = c(0.5,0.5,0.5,0.5), rho_T = 0.5, rho_E = 0.5, rho_C = 0.2,
K=4, m=2000, pcausal=0.1, blocksize=100,
gamma=1, samplesize=700, h2_base=0.3, h2_target=0.3)
re <- mtPRS_FIMEL(dat$base, dat$target, dat$validation, dat$corr, pcut = 1, varcut = 0.8, K = 4, phenotype = "pgx")
hist(re$slPRS_validation)
}

}
\references{
Zhai, S., Guo, B., and Shen, J., 2025. Improving multi-trait polygenic risk score prediction using fine-mapping and ensemble learning.
}
\author{
Song Zhai
}
